Override COMMTIMEOUTS to not wake up with empty data on Windows
The serialport crate sets ReadTotalTimeoutConstant to the specified timeout
value [1], which makes the reader wake up quite often with no payload, and
confuses the tokio wrapper [2] since a 0-byte read usually means EOF.

 [1]: https://gitlab.com/susurrus/serialport-rs/blob/2.3.0/src/windows/com.rs#L224
 [2]: https://github.com/berkowski/mio-serial/issues/7

The MSDN doc [3] is unclear how we can configure a COM port to make it behave
like a socket, but I found an example on StackOverflow [4], so tried it and
it just worked. Still the serial_printline.rs example printed "Read timed out!"
per loop, but the whole "Read would have blocked." mess went away. And my
tokio-based program worked on Windows in the same way as on Linux.

 [3]: https://docs.microsoft.com/en-us/windows/desktop/api/winbase/ns-winbase-_commtimeouts
 [4]: https://stackoverflow.com/a/26007086/10435339

Since I'm not a Windows user, I can't say this is the best and most correct
solution. QtSerialPort appears to set ReadIntervalTimeout to MAXDWORD, but
it didn't go well. I've tested this change on VirtualBox with a pseudo serial
device connected to the host's TCP socket.